Decompose function `Wallet.signTransaction`.
This function had the dual responsibility of assigning change addresses
and signing the resultant transaction.

However, we sometimes need to sign transactions for which change addresses
have already been assigned, or transactions without any change. In the
second case, it's overly restrictive to require that callers provide a
way of making change addresses.

Therefore, it makes sense to decouple these two behaviours.

This change extracts out a new function `assignChangeAddressesAndUpdateDb`,
which handles the responsibility of assigning change addresses and
updating the DB.